+++ CHANGELOG.md
@@ -28,6 +28,9 @@
 - Fix error classes not inheriting from `Exception`
 - Fix an issue with `djangae.storage.CloudStorage` where calling `_open()` or `delete()` wouldn't use the correct bucket
 - Add parameter to control datastore emulator `--data-dir`
+- Remove `environment.queue_name()` function, use `environment.task_queue_name()` instead
+- Update `environment.task_queue_name()` to return `default` if we're in a task and a queue name is not set, otherwise return `None`
+- Update `djangae/tasks/deferred.py` to handle the case where a queue name is not set
 
 ### Bug fixes: